Reorganize code
* Also minor code issues fixes
