+++ setup.py
@@ -13,9 +13,12 @@ ext_modules = [
         [
             'src/cpp/common/py_monero_common.cpp',
             'src/cpp/daemon/py_monero_daemon_model.cpp',
-            'src/cpp/daemon/py_monero_daemon.cpp',
+            'src/cpp/daemon/py_monero_daemon_default.cpp',
+            'src/cpp/daemon/py_monero_daemon_rpc.cpp',
             'src/cpp/wallet/py_monero_wallet_model.cpp',
             'src/cpp/wallet/py_monero_wallet.cpp',
+            'src/cpp/wallet/py_monero_wallet_full.cpp',
+            'src/cpp/wallet/py_monero_wallet_rpc.cpp',
             'src/cpp/py_monero.cpp'
         ],
         include_dirs=[
