+++ src/cpp/daemon/py_monero_daemon_model.cpp
@@ -218,7 +218,7 @@ bool PyMoneroRpcConnection::check_connection(int timeout_ms) {
       throw std::runtime_error("Could not set rpc connection: " + m_uri.get());
     }
 
-    bool connected = m_http_client->connect(std::chrono::milliseconds(timeout_ms));
+    m_http_client->connect(std::chrono::milliseconds(timeout_ms));
     auto start = std::chrono::high_resolution_clock::now();
     PyMoneroJsonRequest request("get_version");
     auto response = send_json_request(request);
